Handle text conflicts
Pull in the extra columns that we need to resolve these conflicts,
dropping them after use. Delay renaming the data column, just to
save looking up the name.

The process of reapplying the value seems to convert to float.
Couldn't find a quick and neat way to stop it doing this, so just
write the floats out to 0 decimal places in the CSV.