Force datatypes to str when reading CSV
We cannot force to other types because user input errors will
make the code fall over here. We expect to clean up such errors in
the output of this script, not in the script itself.

Leave in support for nptype, even though it is not currently used.
It might be useful for more sophisticated auto-resolution.