Make autoresolved store dicts rather than lists
If a single column heading is stored more than once
for a given row (as in the case of years_at_sea_resolver
calling category_resolver) then it will end up listed
twice in the autoresolved field.

Ideally we would use a set for this, but sets are not
ordered so we use dict keys instead.