Look for transcription problems
We look for 00 in strings, and for [] anywhere.

This messes up the 'bad' count -- a [] could be from an
unresolvable cell, or from a resolved cell that contains some
uncertain transcription. So just change the bad counts to report
a minimum rather than a total.

A better solution would be to stop counting bad as we go, and
just base it on cells with square brackets.