Add coverage check, update test input
This might be a bit weird. I've instrumented the code to report
coverage for key code paths.

On the one hand, something similar
could be achieved with careful unit testing and a code coverage
tool, and by doing this I would not clutter the code with
instrumentation.

On the other hand, this gives me a quick way to check that
my coverage is still good, and it serves as inline documentation
in the code that would likely otherwise still be there as a
comment otherwise.

Test inputs updated to cover all the paths with (I hope)
sufficient interesting input to catch problems.