Drop all incomplete rows together
If -u is not set then all incomplete rows should be dropped.
Previously TEXT_T was not dropping these, but they were being
caught later (without the check for -u). DROP_T happened to be
dropping them anyway. Now we drop them all together.

The reason that TEXT_T was not dropping but that DROP_T was is
just that comparison to NaN always yields false and the relevant
logic happened to be arranged such that a false result excluded
a row from entry to DROP_T, but included it for TEXT_T.