Return empty string for garbled 'years at sea' subfield
The idea is that the reducer may be more able to cope with this
as a 'thing to ignore' than when it is presented with random
strings. But the way that we handle this particular field is so
customised that this likely also requires some work in aggregate.py.

Also stop presenting the canonical form as separated by '; ', as
the space gets tricky to track when we are returning blanks on the
right hand side.