Try harder to clean numbers and dates
This includes the earlier idea that the reducers might cope
better with empty strings than with otherwise invalid inputs, but
this idea really needs to be tested either by looking into how the
reducers work, or empirically by looking at the difference it makes
in the results. I am seeing better results now, but there are
confounding factors: I'm running with a more aggressive resolution
parameter for cases where we are choosing the winner by votes, and
there is a lot of cleanup going on in dates.

Testing for these new features is still to do, but I have added a
few test cases for the dates.