Experimenting with the possibility that extraction cleanup can leave us with a blank cell
Some extraction cleanup routines replace wrong inputs with an
empty string, in the hope that this might help the reducers.

Concerned that in this case, ignoring blanks might leave us with
a bunch of cleanup at the end of the project. At the end we
presumably will run a tranche with blanks enabled -- if some of these
blanks could have been reconciled earlier then we would have a lot
of work to do.

Have most of the support for this possibility in place, but not clear
on how the representation works in CSV files.

As it turns out that this possibility seems only to come up for
strings, and as I do not blank strings in the extraction cleanup
script, I can set this aside for now.