Normalise dropdown 14 for workflow 18624 (how disposed)
This is usually "To a/his Ship Cured", but occasionally
"To a Ship Cured".

Added a step to extract.sh to make this always "To a/his..."
Check in the regenerated Task_labels... file
Update quick test golden file accordingly

No independent test for this, but we are testing using a
checked in output from the process, which will have to do.