Drop index from output
We keep the subject id in place, but drop the task numbers, which
are not all that useful.

This is a preparatory step to making the subject id into a link to
the subject image on Zooniverse. While there might be a way to
manipulate the key itself, it seems safer to add a column, and
certainly much quicker to work out.