The numbering scheme changed at some point
Originally volunteers were asked to treat months as 0.08 of a year.
As a result, we rounded all decimal values to multiples of 0.08.

At some point, this was changed to a scheme specifying a particular
decimal for each number of months.

IIUC, we have data of both types. So I think the safest solution
that doesn't require excessive thinking is:

1) If the decimal is one of the numbers prescribed from the new
   scheme, return that (includes 0)
2) Otherwise:
2a) Round to the nearest multiple of 0.08, as before
2b) Convert that multiple to the new scheme with a lookup table

Added some extra tests to cover this.

Also removed the possibility of 0.96 and greater -- this would be
12 months, so surely should be a year? Put in an Exception so I
will catch this if it comes up in the real data.