Generate tranche info
This produces a 'views' file which records the number of
classifications that have come out of each workflow and
also a 'complete' field which is true if every entry in a
given row has at least 3 classifications.

This is complicated a bit by 'port sailed out of' not
applying to volume 1. We just ignore this column if the
current volume is vol 1.

Manipulated testinput/ to cover the testing, as follows:

1)
* Remove 'port sailed out of' row two rows in vol 1 subject 44121901
* Complete 'years at sea' for one of these rows, and also for the fourth row

Years at sea was the only missing classification in these rows.

The effect of this is to test that, both when 'port sailed out of'
is entirely absent, and when it has been removed by the script ignoring
'port sailed out of' classifications for thios workflow, that the row is
considered complete.

2)
* Complete 4 more rows in vol 1 subject 44121901 by adding 'years at sea' entries
* Rearrange votes for 'how disposed of' for the same rows so that:
** One row has one vote for each of 3 categories (still 'complete')
** One row has 2 votes for a category (incomplete)
** One row has 1 votes for a category (incomplete)
** One row is removed entirely (i.e. no votes, incomplete)

This tests various dropdown cases (as opposed to the text cases above).