Test that port sailed out of is required outside of vol 1
Add enough entries to complete 3 rows in vol 2.

Remove 'port sailed out of' from two of them -- in one case
by giving it insufficient classifications, in the other by
removing the row entirely.

Confirm that only the complete row has 'complete' = True in
the views file.

Complete an entry in volume 2, preparatory to confirming that years at sea is a requirement to be True in views in vol 2

Complete a row in volume 2 (prep for demonstrating that years at sea is required in vol 2)

Test that "years at sea" is required outside of vol 1