Add second round of testing for tranching
We manipulate the input files so that they look as they would if
the views file from the previous round had been used at extraction
time.

The result should be a main output that is exactly the same except
for missing those particular rows, and a views file that is
entirely identical.