Script to install Python deps via pip
Removes the upper limits on the dependencies of
panoptes_aggregation, which is risky as we have not re-tested
panoptes_aggregation with these particular packages.

Without this approach I am unable to install in a virtualenv due
to what seems to be an ABI incompatibility between different
numpy versions. Google phrases such as
numpy.ndarray size changed, may indicate binary incompatibility
for more.

There is also this thread on Zooniverse Talk:
https://www.zooniverse.org/talk/1322/2415041