Rewrite extract.sh as a Python script
Overdue really, the shell was getting quite nasty. Will make it
easier to extend the functionality to deal with multiple versions
of the same workflow.

As a side-effect, we finally get to use workflow.yaml to drive
this, which hopefully has got rid of all duplicate workflow
definitions.

Also add an 'export' field to the workflow yaml, to replace
the array of CSV files in the old shell script.