Deal with different versions of the same workflow
A couple of HMS NHS workflows changed in minor ways during
phase 1. Generalise the script to be able to process all
relevant versions of the workflow, concatenate them together into
one and reduce them, checking that the config for the reduction
should be valid.

Do not yet actually add further versions to the yaml file.