Extend identity checking
The idea was to confirm that all outputs of panoptes config are
the same for all versions of a given workflow.

However, extraction files are different because they necessarily
contain the version.

And Task files can be different in hex strings, which are hopefully
an unimportant detail -- added a FIXME to remind me to confirm
this.

While it would be good to confirm that these files are identical
modulo expected changes, this is a lot of extra work with
limited value for what we need these scripts to do.