Count number of views accurately
The text reducer does not count empty strings, but sometimes
users enter them. (It probably makes sense for the reducer not
to consider empty strings while reducing.)

So instead we count classification rows in the extractions file.

We cannot use retirement data in the -subjects file because we
cannot guarantee that the data will still be in that file at
any given time. It might be more efficient to use the retirement
metadata in the exports file, but I didn't think of that until
after I had implemented it this way.

We do not want to risk messing up our threshold calculation, which
relies upon numbers related by the reducers count, so we keep this
updated count separate from the "main" dataframe.