Include multiple classifications by one user on one subject in the reduction
Sometimes a given user has classified the same record more than
once. We are now counting all of these classifications when working
out the number of views. We now also use them in the reduction,
rather than just taking the last classification. Previously I
was taking the last classification on the principle that later
classifications might be fixing mistakes. But I have seen a case
where a later classification just blanked everything, which seems
more likely to be a mistake. There may be something more
sophisticated or "right" to do here, but that's another project.

As it does seem undesirable to have multiple classifications from
a single user, also create some basic reporting on this (but only
for the text reductions, as no-one actually asked for this feature).