Move view count up and use it to decide which text classifications to keep
Moving it up here lets us get rid of the calculated count column
and just track this in views instead.

It also allows us to use the same logic to drop rows for both
text and dropdowns.

A consequence is that it no longer makes sense to dropna rows
based on data.number_views, as we are counting views based on
entries in the extraction file, rather than doing a numerical
comparison on a field that might be na. So na values can be
passed through even when arg.unfinished is False. So we update
the logic to deal with that, modifying the condition to a form
that I think makes a bit more sense (check number_views, then
essentially assert that another field is null in this case).

This does change the output a bit -- now the view count lets
through classifications that we previously discarded, with values
based on an "insufficient" number of non-blank entries. This
(at a minimum) means that text entered by one person, where two
people entered nothing, will now get through as the 100% certain
classification. This might be desirable, especially if we later
create a means of knowing what the "undercounted" fields are.