Makefile converts blanks to 00 in both golden and output
Also update differ.sh blank detection to match.

Makes the way that we say "blank cell" uniform. But this does
imply that we will apply a similar transformation in aggregate.py
(either make 00 into blank, or blank into 00), which in turn
means that we are treating "implicit blank" and "explicit blank"
the same way.