Initial attempt at handling dates better
The warnings help with understanding the error patterns in the
data, which we can then attempt to fix up. There is more to do
here, e.g. identifying which phase we are in will help with
century-disambiguation, we can handle dates with misplaced separators
(e.g. 2-109-1830 is likely 21-09-1830), dates in reverse order (1830-09-21).

If we wanted to get more ambitious, we could look for contributors
who consistently use a particular incorrect date pattern (e.g. mm-dd-yy).

And it would make a lot of sense to reconcile date, month and year
separately -- one quick-and-dirty way to do this would be to
generate 3 extractions files per date field.