Handle 0-fields in dates
The date parser that we currently rely upon does not cope well
with zero-fields inside a date.

If it looks to the cleaner like any part of the date contains a
zero-field then it just formats it as best it can and returns it.
The aggregator checks all candidates for a date -- even when
autoresolved -- and returns unresolvable if any candidate seems
to contain either a zero-field or a weird number of fields.

This could all be done better: for initial thoughts see
https://github.com/nationalarchives/hms-nhs-scripts/issues/11