Improve post-reconciliation "transcription uncertainty" detection
Add more patterns to post-reconciliation detection of transcriber's
uncertainty indications. The goal here is not to catch every last
case, as the more we try to catch, the more noise we get. But we
can at least catch a lot of cases. A very brief, incomplete
eyeballing of the results suggests that the current set of
patterns is not adding loads of false positives.

Also, the --uncertainty flag does pre-reconciliation detection.
Extend its help text to make this clear, and to give some idea of
the implications.