Fix task matching in strip_processed
The views file has numerical task identifiers, while the
extractions file preceeds the numbers with a T. To fix, remove
the prefix in the extractions file and put it back later. This
is hopefully quicker, as it allows the join to be based on
numbers rather than strings, and it also simplifies the sorting
a bit.