Record the two ways of dropping rows/pages separately
There are two places where we remove things for being incomplete.
These can be found by searching the file for "args.unfinished".

The first is at line 475 in this commit. Here we are processing
the individual reduction files (==columns) and we drop all entries
in the column with insufficient views to be counted. This ensures
that these inputs will have no effect upon later processing.

The second is at line 675 in this commit. Here we are removing
pages which contain any incomplete or entirely missing rows.

Previously we were not recording things removed in the first case
unless they also had a visible effect in the second case -- this
would only happen if other data made a sufficiently-viewed entry
for the same page. The lack of this output caused me to miss a
problem, so it does genuinely matter.

Rather than try to reconcile these into a single 'removed.csv', we
rather dump them into separate files. This ensures that we can see
what was dropped for either reason without complicating the code
or the output files.