Stop dropping completed subjects from joined_views
This seems pointless, given that this df contains a complete
column that tells me which rows are complete and which are not --
I'm just removing information that can help me to understand the
data that I have processed.

There is a chance that this somehow breaks the tranching, but
that is anyway an incomplete and probably un-needed feature.