Update to latest Panoptes, record pip deps
Include a minimal extract.py tweak to work with newer panoptes,
a script for comparing parts of the extract.py output, and a
tweak to clean_extraction.py to sort part of its output to
ensure that it remains easy to compare.

The newer panoptes seem quite a bit quicker and does not change
the output.

We should now be able to use the newer panoptes to handle multiple
versions, but that is a problem for later.

Note that I had to force joblib to a lower version -- see
https://stackoverflow.com/questions/73830225/init-got-an-unexpected-keyword-argument-cachedir-when-importing-top2vec

Moved an old 'pip deps installer' to misc_scripts and added a
comment noting that it no longer seems to be needed.