Add facilities to filter (at extract.py-time) for particular volumes
This lets us deal with cases where workflow version does not change,
but project phase does. Volume is also a more straightforward way
of identifying project phase.

Update workflow.yaml to apply this volume limiting to phase1. Note
that this will (correctly) cause classifications for a pair of subjects
(subject ids 44444750 and 44444752) to get dropped from the
admission number workflow, as those subjects are in later volumes
but appear to have been classified once with the older workflow
version.

Because we have added new per-workflow fields, create a new level
of hierarchy in the YAML file to record the information for all
individual workflows.

We may still need other means of identifying project phase, as
test classifications may exist under some particular version number,
for example.