Add phase2
Descriptions of the workflows themselves are added with no
versions: we just pull everything in for now and pick the
cases in the correct volume range.

clean_extraction has rules added for each new workflow. We
do not add any special cleaning for any of the new workflows,
just use the generic option for the workflow type.

We now must normalize / in names to _ when writing out files.

We identify number workflows from the YAML file, rather than
relying upon a hardcoded list that only applies to phase1.

We only do port_sailed_out_of special steps when volume 1 is
present in the dataset.

Handling mismatch between the current_views and the actual
reductions index is now much more involved. This is because
the reductions index now sometimes contains null entries
where no classifications have been given for some tasks in
a workflow. I believe that these can safely be ignored:
see comments in the code for more details.