Add notes about versions
config_file_comparisons_phase2.ods
==================================

This information gathered from running extract.py and looking at
the output to see which versions differ, then using diff tools to
understand differences in the extractions config.

The 'Count' gives the number of rows corresponding to that workflow
and version in the exports file. This should be the number of
classifications for that workflow and version.

Values in the Labels, Extraction and Reduction column are Base,
Matches or Differ. Base is the config that was used as a baseline
for comparison. Matches means that a given version matched that
config. Differs (highlighted in red) means that is does not. For
Extractions, comments have been added to describe the difference.
This always amounts to 8 more slots being made avaiable for phase2,
except in a couple of cases which I have flagged with bold text in
column H.

Versions highlighted in green do not differ from the "base" version.

versions_per_volume.csv
=======================

I forget how I generated this, but it appears to be a snapshot of
volumes and the workflow versions that have generated
classifications for them.