Dynamic fields for mimsify.py
mimsify.py was relying upon a hardcoded list of field names to
work with. We now load the fields from the appropriate phase in
workflow.yaml instead, which means that phase1 and phase2 can
both work.

Had to rework the CLI a little to facilitate this. Happily, this
made it a little bit more consistent with the other scripts.