Blank out "explicit blank" text
The documentation asks for "No Row" but I have noticed a few
variations on this.

This can sometimes result in bad transcriptions for a cell or two
on a line getting through -- previously they would have competed
with No Row and may well have lost to it. But the benefit of
blanking it here, rather than dealing with this entirely in
aggregate.py, is that we avoid problems with transcriptions and
No Row entries getting merged together, when both have been
entered for a single cell.

It's a bit swings and roundabouts -- perhaps this approach is
better, perhaps it is better to handle it all in aggregate.py.
Probably the right thing to do is to think of a more sophisticated
way of handling explicit blanks.