Drop entirely blank rows
Rows consisting entirely of explicit blanks, or rows that have
just been skipped, will contain empty fields and/or zero values.
If a whole row contains such value then we assume that it is blank
in the original text and drop it.

We drop it from the views file too -- at this point we cannot be
sure whether it has been explicitly classed as empty, or just
skipped. The consequence of dropping it from the views file should
be to ensure that it will always be re-read when we run a new tranche.